[Cherry-pick][Enhancement] optimize the re-balance strategy for backend scan range (#28468) (#31683)

+++ fe/fe-core/src/test/java/com/starrocks/qe/HDFSBackendSelectorTest.java
@@ -52,7 +52,7 @@ public class HDFSBackendSelectorTest {
     final int computeNodePort = 9030;
     final String hostFormat = "Host%02d";
 
-    private List<TScanRangeLocations> createScanRanges(int number, int size) {
+    private List<TScanRangeLocations> createScanRanges(long number, long size) {
         List<TScanRangeLocations> ans = new ArrayList<>();
 
         for (int i = 0; i < number; i++) {
